Fixes Gift Attachments, Conserv Attachments, made schema change for TableID type from Byte to Short.
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@11015 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/datamodel/Attachment.java
@@ -90,7 +90,7 @@ public class Attachment extends DataModelObjBase implements Serializable
     protected String                  remarks;
     protected String                  attachmentLocation;
     protected String                  guid;
-    protected Byte                    tableID;
+    protected Short                   tableID;
     protected Byte                    visibility;
     protected SpecifyUser             visibilitySetBy;
     protected Set<AttachmentMetadata> metadata;
@@ -375,7 +375,7 @@ public class Attachment extends DataModelObjBase implements Serializable
      * @return the tableID
      */
     @Column(name = "TableID", unique = false, nullable = false, insertable = true, updatable = true)
-    public Byte getTableID()
+    public Short getTableID()
     {
         return tableID;
     }
@@ -383,7 +383,7 @@ public class Attachment extends DataModelObjBase implements Serializable
     /**
      * @param tableID the tableID to set
      */
-    public void setTableID(Byte tableID)
+    public void setTableID(Short tableID)
     {
         this.tableID = tableID;
     }
@@ -395,7 +395,7 @@ public class Attachment extends DataModelObjBase implements Serializable
     {
         AppContextMgr acm = AppContextMgr.getInstance();
 
-        this.tableID = (byte)tableIDArg;
+        this.tableID = (short)tableIDArg;
         
         if (tableIDArg == Accession.getClassTableId())
         {
